@@ -2,7 +2,8 @@
   "name": "applicationinsights",
   "license": "MIT",
   "bugs": "https://github.com/Microsoft/ApplicationInsights-node.js/issues",
-  "version": "1.8.0",
+  "version": "2.0.0",
+  "private": true,
   "description": "Microsoft Application Insights module for Node.js",
   "repository": {
     "type": "git",
@@ -38,31 +39,31 @@
     "build": "npm run build:deps && npm run build:compile",
     "build:deps": "npm update --dev",
     "build:compile": "tsc --project ./tsconfig.json",
-    "prepare": "npm run build:compile",
     "prepublishOnly": "npm run build",
-    "pretest": "npm run build",
-    "test": "npm run test:ts && npm run test:js",
-    "test:debug": "mocha ./out/Tests --inspect-brk --recursive --no-exit",
-    "test:ts": "mocha ./out/Tests --recursive --no-exit",
-    "test:js": "mocha ./Tests/js --recursive --no-exit",
+    "test": "ts-mocha ./Tests/**/*.ts --recursive --no-exit",
+    "test:debug": "ts-mocha ./Tests/**/*.ts --inspect-brk --recursive --no-exit",
     "functionaltest": "npm run build && npm pack && node --use_strict ./Tests/FunctionalTests/RunFunctionalTests.js",
     "backcompattest": "npm run build && npm pack && node --use_strict ./Tests/BackCompatibility/RunBackCompatTests.js"
   },
   "devDependencies": {
-    "@types/cls-hooked": "^4.3.0",
-    "@types/mocha": "2.2.48",
-    "@types/node": "4.2.4",
-    "@types/sinon": "2.1.2",
-    "applicationinsights-native-metrics": "0.0.5",
-    "mocha": "3.5.2",
-    "node-mocks-http": "1.2.3",
-    "sinon": "1.17.6",
-    "typescript": "2.4.2"
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^14.0.13",
+    "@types/sinon": "^9.0.4",
+    "mocha": "^8.0.1",
+    "node-mocks-http": "^1.8.1",
+    "sinon": "^9.0.2",
+    "ts-mocha": "^7.0.0",
+    "typescript": "^3.9.5"
   },
   "dependencies": {
-    "cls-hooked": "^4.2.2",
-    "continuation-local-storage": "^3.2.1",
-    "diagnostic-channel": "0.3.1",
-    "diagnostic-channel-publishers": "0.4.0"
+    "@azure/monitor-opentelemetry-exporter": "^1.0.0-preview.2",
+    "@opentelemetry/api": "^0.9.0",
+    "@opentelemetry/context-async-hooks": "^0.9.0",
+    "@opentelemetry/core": "^0.9.0",
+    "@opentelemetry/node": "^0.9.0",
+    "@opentelemetry/plugins-node-all": "^0.9.0",
+    "@opentelemetry/tracing": "^0.9.0",
+    "diagnostic-channel": "^0.3.1",
+    "diagnostic-channel-publishers": "^0.4.0"
   }
 }
