@@ -1,9 +1,8 @@
 import Contracts = require("../Declarations/Contracts")
 import Util = require("./Util")
 import Config = require("./Config");
-import Context = require("./Context");
-import { CorrelationContextManager } from "../AutoCollection/CorrelationContextManager";
-
+import Context = require("./SdkContext");
+import * as opentelemetry from "@opentelemetry/api";
 
 /**
  * Manages the logic of creating envelopes from Telemetry objects
@@ -268,7 +267,10 @@ class EnvelopeFactory {
     }
 
     private static getTags(context: Context, tagOverrides?: { [key: string]: string; }) {
-        var correlationContext = CorrelationContextManager.getCurrentContext();
+        var correlationContext = opentelemetry.trace
+            .getTracer('applicationinsights')
+            .getCurrentSpan()
+            ?.context();
 
         // Make a copy of context tags so we don't alter the actual object
         // Also perform tag overriding
@@ -287,9 +289,12 @@ class EnvelopeFactory {
 
         // Fill in internally-populated values if not already set
         if (correlationContext) {
-            newTags[context.keys.operationId] = newTags[context.keys.operationId] || correlationContext.operation.id;
-            newTags[context.keys.operationName] = newTags[context.keys.operationName] || correlationContext.operation.name;
-            newTags[context.keys.operationParentId] = newTags[context.keys.operationParentId] || correlationContext.operation.parentId;
+            newTags[context.keys.operationId] = newTags[context.keys.operationId] || correlationContext.traceId;
+
+            // @todo
+            // newTags[context.keys.operationName] = newTags[context.keys.operationName] || correlationContext.traceId;
+
+            newTags[context.keys.operationParentId] = newTags[context.keys.operationParentId] || correlationContext.spanId;
         }
 
         return newTags;
