@@ -25,7 +25,7 @@ describe("AutoCollection/Performance", () => {
 
     describe("#trackNetwork()", () => {
         it("should not produce incorrect metrics because of multiple instances of Performance class", (done) => {
-            const setIntervalStub = sinon.stub(global, "setInterval", () => ({ unref: () => {}}));
+            const setIntervalStub = sinon.stub(global, "setInterval").callsFake(() => ({ unref: () => {}}) as number & object);
             const clearIntervalSpy = sinon.spy(global, "clearInterval");
             const appInsights = AppInsights.setup("key").setAutoCollectPerformance(false).start();
             const performance1 = new Performance(new TelemetryClient("key"), 1234, false);
@@ -35,34 +35,34 @@ describe("AutoCollection/Performance", () => {
             Performance.INSTANCE.enable(true);
             const stub1 = sinon.stub(performance1["_client"], "trackMetric");
             const stub2 = sinon.stub(performance2["_client"], "trackMetric");
-
             Performance.countRequest(1000, true);
             Performance.countRequest(2000, true);
-            performance1["_trackNetwork"]();
-            performance2["_trackNetwork"]();
-            Performance.countRequest(5000, true);
-            const prev1 = performance1["_lastIntervalRequestExecutionTime"];
-            const prev2 = performance2["_lastIntervalRequestExecutionTime"];
-            assert.deepEqual(prev1, prev2);
-            assert.deepEqual(prev1, 1000 + 2000);
 
             // Add to end of event loop
             setTimeout(() => {
+                performance1["_trackNetwork"]();
+                performance2["_trackNetwork"]();
+                Performance.countRequest(5000, true);
+                const prev1 = performance1["_lastIntervalRequestExecutionTime"];
+                const prev2 = performance2["_lastIntervalRequestExecutionTime"];
+                assert.deepEqual(prev1, prev2);
+                assert.deepEqual(prev1, 1000 + 2000);
+
                 assert.equal(Performance["_intervalRequestExecutionTime"], 1000 + 2000 + 5000);
-                assert.equal(stub1.callCount, 2, "calls trackMetric for the 2 standard metrics");
-                assert.equal(stub2.callCount, 3, "calls trackMetric for the 3 live metric counters");
+                assert.equal(stub1.callCount, 2, "(primary) calls trackMetric for the 2 standard metrics");
+                assert.equal(stub2.callCount, 3, "(primary) calls trackMetric for the 3 live metric counters");
                 assert.equal(stub2.args[1][0].value, stub1.args[1][0].value);
                 assert.equal(stub1.args[1][0].value, (1000 + 2000) / 2, "request duration average should be 1500");
 
-                stub1.reset();
-                stub2.reset();
+                stub1.resetHistory();
+                stub2.resetHistory();
 
                 setTimeout(() => {
                     // need to wait at least 1 ms so trackNetwork has valid elapsedMs value
                     performance1["_trackNetwork"]();
                     performance2["_trackNetwork"]();
-                    assert.equal(stub1.callCount, 2, "calls trackMetric for the 2 standard metrics");
-                    assert.equal(stub2.callCount, 3, "calls trackMetric for the 3 live metric counters");
+                    assert.equal(stub1.callCount, 2, "(secondary) calls trackMetric for the 2 standard metrics");
+                    assert.equal(stub2.callCount, 3, "(secondary) calls trackMetric for the 3 live metric counters");
                     assert.equal(stub2.args[1][0].value, stub1.args[1][0].value);
                     assert.equal(stub1.args[1][0].value, (5000) / 1, "request duration average should be 5000");
 
@@ -75,8 +75,8 @@ describe("AutoCollection/Performance", () => {
                     setIntervalStub.restore();
                     clearIntervalSpy.restore();
                     done();
-                }, 100);
-            }, 100);
+                }, 1);
+            }, 1);
         });
     });
 });
